requirements.txt
numba      
matplotlib  
fire     
protobuf  
opencv-python 
opencv-contrib-python  
pybind11                                      
easydict      
open3d-python 
terminaltables  
pytest-runner    
addict           
pycocotools      
imagecorruptions  
objgraph       
cachetools
descartes
jupyter
matplotlib
motmetrics<=1.1.3
numpy
pandas>=0.24
# Pillow<=6.2.1  # Latest Pillow is incompatible with current torchvision, https://github.com/pytorch/vision/issues/1712
Pillow==9.0.1  
pyquaternion>=0.9.5
scikit-learn
scikit-image
scipy  
Shapely
tqdm
pyyaml
requests
# nuscenes-devkit==1.1.6
cmake==3.18.4.post1
Cython

# torch_scatter==2.0.7 # err for 2.0.9 